+++ src/main/java/com/example/kinobackend/theater/ITheaterRepository.java
@@ -1,6 +1,5 @@
-package com.example.kinobackend.repository;
+package com.example.kinobackend.theater;
 
-import com.example.kinobackend.model.Theater;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
